% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathprog_example.R
\docType{data}
\name{assignment_example}
\alias{assignment_example}
\title{Formulation of the assignment problem in MathProg}
\format{
An object of class \code{character} of length 1.
}
\usage{
assignment_example
}
\description{
A formulation of the assignment problen using MathProg, and an instance of size eigth from Christofides.
}
\examples{
writeLines(assignment_example)


}
\keyword{datasets}
