% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plem_solver.R
\name{plem_solver}
\alias{plem_solver}
\title{Solving a PLEM model using Rglpk}
\usage{
plem_solver(variables = list(), verbose = T, want_0 = F, math_txt = F)
}
\arguments{
\item{variables}{A named list. The element named lp.code is the model to solve.}

\item{verbose}{A flag indicating if we want a verbose output.}

\item{want_0}{Do you want the values equal to zero of the solution? (\code{FALSE} by default).}

\item{math_txt}{Do you want the MathProg files to be tested elsewhere? (\code{FALSE} by default).}
}
\value{
A list including:
\itemize{
\item \code{sol} the optimum value of the objective function.
\item \code{vars} a data frame with the optimal solution.
}
}
\description{
A wrapper to load a PLEM model and solve it using Rglpk.
}
\examples{
plem_solver(list(lp.code = mathprog_example))

}
