% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathprog_example.R
\docType{data}
\name{mathprog_example}
\alias{mathprog_example}
\title{An example of linear programming in MathProg}
\format{
An object of class \code{character} of length 1.
}
\usage{
mathprog_example
}
\description{
A Small Example of Linear Programming in MathProg
}
\examples{
writeLines(mathprog_example)


}
\keyword{datasets}
