% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plem_solver.R
\name{plem_solver}
\alias{plem_solver}
\title{Solving a PLEM model using Rglpk}
\usage{
plem_solver(variables = list(), verbose = T, want_0 = F, math_txt = F)
}
\arguments{
\item{variables}{A named list. The element named lp.code is the model to solve.}

\item{verbose}{A flag indicating if we want a verbose output.}

\item{want_0}{A flag variable.}

\item{math_txt}{A flag variable.}
}
\value{
The solution of the model, as delivered by Rglpk.
}
\description{
A wrapper to load a PLEM model and solve it using Rglpk.
}
\examples{
plem_solver(list(lp.code = mathprog_example))

}
